streamlit-app/requirements.txt
streamlit
data-preprocessors
bangla-postaggers